Update on "[quant][graphmode] Fp16 quant support - remove activation cast op"
Summary:
In eager mode there is no cast operator for the activation tensor in the fbgemm fp16 operator
Remove it from graph mode. This makes numerics match between debug model and final quantized model.

Test Plan:
python test/test_quantization.py test_linear_dynamic_fp16
Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
