+++ aten/src/ATen/native/native_functions.yaml
@@ -3615,6 +3615,10 @@
   use_c10_dispatcher: full
   variants: function
 
+- func: _saturate_weight_to_fp16(Tensor weight) -> Tensor
+  use_c10_dispatcher: full
+  variants: function
+
 # to(Device) must not exist because all constructors of Device also works for
 # TensorOptions. Otherwise, an ambiguity error is thrown.
 # See NOTE [ TensorOptions Constructors ].
